Finalizing file locking and relocating to new Caching feature

+++ Libraries/Caching/LockFile.cs
@@ -4,14 +4,17 @@ using System.Linq;
 using System.Text;
 using Orchard.Environment.Extensions;
 using Orchard.FileSystems.Media;
+using System.IO;
 
-namespace Piedone.HelpfulLibraries.Tasks
+namespace Piedone.HelpfulLibraries.Caching
 {
-    [OrchardFeature("Piedone.HelpfulLibraries.Tasks")]
+    [OrchardFeature("Piedone.HelpfulLibraries.Caching")]
     public class LockFile : ILockFile
     {
         private readonly IStorageProvider _storageProvider;
         private string _name;
+        private bool isDisposed = false;
+        private bool isAcquired = false;
 
         private const string _folder = "HelpfulLibraries/Tasks/LockFiles/";
 
@@ -23,27 +26,22 @@ namespace Piedone.HelpfulLibraries.Tasks
         public bool TryAcquire(string name)
         {
             _name = name;
+            isAcquired = true;
 
             // Here really should be a file existence check, but it's not possible currently:
             // http://orchard.codeplex.com/workitem/18279
-            try
+            using (var stream = new MemoryStream())
             {
-                _storageProvider.CreateFile(MakeFilePath(name));
-                return true;
+                return _storageProvider.TrySaveStream(MakeFilePath(name), stream);
             }
-            catch (ArgumentException)
-            {
-                // The file exists
-            }
-
-            return false;
         }
 
         public void Dispose()
         {
-            if (String.IsNullOrEmpty(_name)) return;
+            if (String.IsNullOrEmpty(_name) || isDisposed || !isAcquired) return;
 
-            // Could throw exception...
+            isDisposed = true;
+            // Could throw exception e.g. if the file was deleted. This should not happen.
             _storageProvider.DeleteFile(MakeFilePath(_name));
         }
